backend/pyproject.toml
# backend/pyproject.toml — зависимости и инструменты backend
# Назначение: управление зависимостями FastAPI (Poetry)

[tool.poetry]
name = "tamagotchi-backend"
version = "0.1.0"
description = "AI tamagotchi backend"
authors = ["Your Name <you@example.com>"]
package-mode = false  # (я добавил)

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}  # (я добавил)
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic-settings = "^2.2.0"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"  # (я добавил)
mypy = "^1.10.0"  # (я добавил)

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
